refactor: refactor trace cli
